ats/requirements.txt
pinecone
python-dotenv
pdfplumber
groq
langchain-pinecone